refactor(portfolio): ui interaction handler
